+++ core/src/test/kotlin/net/corda/core/flows/FastThreadLocalTest.kt
@@ -23,7 +23,6 @@ import net.corda.core.internal.rootCause
 import net.corda.core.utilities.getOrThrow
 import org.assertj.core.api.Assertions.catchThrowable
 import org.hamcrest.Matchers.lessThanOrEqualTo
-import org.junit.After
 import org.junit.Assert.assertThat
 import org.junit.Test
 import java.util.*
@@ -41,30 +40,9 @@ class FastThreadLocalTest {
     }
 
     private val expensiveObjCount = AtomicInteger()
-    private lateinit var pool: ExecutorService
-    private lateinit var scheduler: FiberExecutorScheduler
-    private fun init(threadCount: Int, threadImpl: (Runnable) -> Thread) {
-        pool = Executors.newFixedThreadPool(threadCount, threadImpl)
-        scheduler = FiberExecutorScheduler(null, pool)
-    }
-
-    @After
-    fun poolShutdown() = try {
-        pool.shutdown()
-    } catch (e: UninitializedPropertyAccessException) {
-        // Do nothing.
-    }
-
-    @After
-    fun schedulerShutdown() = try {
-        scheduler.shutdown()
-    } catch (e: UninitializedPropertyAccessException) {
-        // Do nothing.
-    }
 
     @Test
-    fun `ThreadLocal with plain old Thread is fiber-local`() {
-        init(3, ::Thread)
+    fun `ThreadLocal with plain old Thread is fiber-local`() = scheduled(3, ::Thread) {
         val threadLocal = object : ThreadLocal<ExpensiveObj>() {
             override fun initialValue() = ExpensiveObj()
         }
@@ -73,8 +51,7 @@ class FastThreadLocalTest {
     }
 
     @Test
-    fun `ThreadLocal with FastThreadLocalThread is fiber-local`() {
-        init(3, ::FastThreadLocalThread)
+    fun `ThreadLocal with FastThreadLocalThread is fiber-local`() = scheduled(3, ::FastThreadLocalThread) {
         val threadLocal = object : ThreadLocal<ExpensiveObj>() {
             override fun initialValue() = ExpensiveObj()
         }
@@ -83,8 +60,7 @@ class FastThreadLocalTest {
     }
 
     @Test
-    fun `FastThreadLocal with plain old Thread is fiber-local`() {
-        init(3, ::Thread)
+    fun `FastThreadLocal with plain old Thread is fiber-local`() = scheduled(3, ::Thread) {
         val threadLocal = object : FastThreadLocal<ExpensiveObj>() {
             override fun initialValue() = ExpensiveObj()
         }
@@ -93,8 +69,8 @@ class FastThreadLocalTest {
     }
 
     @Test
-    fun `FastThreadLocal with FastThreadLocalThread is not fiber-local`() {
-        init(3, ::FastThreadLocalThread)
+    fun `FastThreadLocal with FastThreadLocalThread is not fiber-local`() =
+            scheduled(3, ::FastThreadLocalThread) {
         val threadLocal = object : FastThreadLocal<ExpensiveObj>() {
             override fun initialValue() = ExpensiveObj()
         }
@@ -103,7 +79,7 @@ class FastThreadLocalTest {
     }
 
     /** @return the number of times a different expensive object was obtained post-suspend. */
-    private fun runFibers(fiberCount: Int, threadLocalGet: () -> ExpensiveObj): Int {
+    private fun SchedulerContext.runFibers(fiberCount: Int, threadLocalGet: () -> ExpensiveObj): Int {
         val fibers = (0 until fiberCount).map { Fiber(scheduler, FiberTask(threadLocalGet)) }
         val startedFibers = fibers.map { it.start() }
         return startedFibers.map { it.get() }.count { it }
@@ -137,8 +113,7 @@ class FastThreadLocalTest {
         }::get)
     }
 
-    private fun contentIsNotSerialized(threadLocalGet: () -> UnserializableObj) {
-        init(1, ::FastThreadLocalThread)
+    private fun contentIsNotSerialized(threadLocalGet: () -> UnserializableObj) = scheduled(1, ::FastThreadLocalThread) {
         // Use false like AbstractKryoSerializationScheme, the default of true doesn't work at all:
         val serializer = Fiber.getFiberSerializer(false)
         val returnValue = UUID.randomUUID()
@@ -170,6 +145,23 @@ class FastThreadLocalTest {
             // In retainObj false case, check this doesn't attempt to serialize fields of currentThread:
             Fiber.parkAndSerialize { fiber, _ -> bytesFuture.capture { serializer.write(fiber) } }
             return returnValue
+        }
+    }
+
+    private data class SchedulerContext(private val pool: ExecutorService, val scheduler: FiberExecutorScheduler) {
+        fun shutdown() {
+            pool.shutdown()
+            scheduler.shutdown()
+        }
+    }
+
+    private fun scheduled(threadCount: Int, threadImpl: (Runnable) -> Thread, test: SchedulerContext.() -> Unit) {
+        val pool = Executors.newFixedThreadPool(threadCount, threadImpl)
+        val ctx = SchedulerContext(pool, FiberExecutorScheduler(null, pool))
+        try {
+            ctx.test()
+        } finally {
+            ctx.shutdown()
         }
     }
 }