+++ core/src/test/kotlin/net/corda/core/utilities/NetworkHostAndPortTest.kt
@@ -10,9 +10,9 @@
 
 package net.corda.core.utilities
 
+import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.Test
 import kotlin.test.assertEquals
-import org.assertj.core.api.Assertions.assertThatThrownBy
 
 class NetworkHostAndPortTest {
     /**