+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/network/NetworkBootstrapper.kt
@@ -360,7 +360,7 @@ class NetworkBootstrapper
                     notaries = notaryInfos,
                     modifiedTime = Instant.now(),
                     maxMessageSize = 10485760,
-                    maxTransactionSize = Int.MAX_VALUE,
+                    maxTransactionSize = 10485760,
                     whitelistedContractImplementations = whitelist,
                     epoch = 1,
                     eventHorizon = 30.days