+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/internal/network/NetworkMapServer.kt
@@ -46,7 +46,7 @@ import javax.ws.rs.core.Response.ok
 import javax.ws.rs.core.Response.status
 
 class NetworkMapServer(private val pollInterval: Duration,
-                       hostAndPort: NetworkHostAndPort,
+                       hostAndPort: NetworkHostAndPort = NetworkHostAndPort("localhost", 0),
                        private val networkMapCertAndKeyPair: CertificateAndKeyPair = createDevNetworkMapCa(),
                        private val myHostNameValue: String = "test.host.name",
                        vararg additionalServices: Any) : Closeable {